+++ tests/test_endpoints.py
@@ -817,6 +817,41 @@ class TestHotWaterTankSection:
                 ssl=False,
             )
 
+    @pytest.mark.asyncio
+    async def test_get_fresh_water_module_temperature(self) -> None:
+        """Test get fresh water module temperature."""
+        with aioresponses() as mock_keenergy_api:
+            mock_keenergy_api.post(
+                "http://mocked-host/var/readWriteVars",
+                payload=[
+                    {
+                        "name": "APPL.CtrlAppl.sParam.hotWaterTank[0].FreshWater.freshWaterTemp.values.actValue",
+                        "attributes": {
+                            "formatId": "fmt3p2",
+                            "longText": "FWM temp.",
+                            "unitId": "Temp",
+                            "upperLimit": "100",
+                            "lowerLimit": "0",
+                        },
+                        "value": "30",
+                    }
+                ],
+                headers={"Content-Type": "application/json;charset=utf-8"},
+            )
+
+            client: KebaKeEnergyAPI = KebaKeEnergyAPI(host="mocked-host")
+            data: float = await client.hot_water_tank.get_fresh_water_module_temperature()
+
+            assert isinstance(data, float)
+            assert data == 30.0  # noqa: PLR2004
+
+            mock_keenergy_api.assert_called_once_with(
+                url="http://mocked-host/var/readWriteVars",
+                data='[{"name": "APPL.CtrlAppl.sParam.hotWaterTank[0].FreshWater.freshWaterTemp.values.actValue", "attr": "1"}]',
+                method="POST",
+                ssl=False,
+            )
+
 
 class TestHeatPumpSection:
     @pytest.mark.asyncio